requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
-e git+https://github.com/openai/CLIP@8cad3a736a833bc4c9b4dd34ef12b52ec0e68856#egg=clip
colorama==0.4.4
einops==0.3.0
filelock==3.0.12
fsspec==2021.7.0
ftfy==6.0.3
future==0.18.2
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
huggingface-hub==0.0.12
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
joblib==1.0.1
kornia==0.5.7
Markdown==3.3.4
multidict==5.1.0
numpy==1.21.1
oauthlib==3.1.1
omegaconf==2.1.0
packaging==21.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==2.4.7
pytorch-lightning==1.4.0
pytorch-ranger==0.1.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
six==1.16.0
-e git+https://github.com/CompVis/taming-transformers@9d17ea64b820f7633ea6b8823e1f78729447cb57#egg=taming_transformers
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tokenizers==0.10.3
torch==1.9.0+cu111
torch-optimizer==0.1.0
torchaudio==0.9.0
torchmetrics==0.4.1
torchvision==0.10.0+cu111
tqdm==4.62.0
transformers==4.9.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
yarl==1.6.3
